moved utility classes and methods into future package
