+++ prestans3/types/model.py
@@ -10,10 +10,9 @@
 """
 from copy import copy
 
-from prestans3.errors import ValidationException, AccessError, ContainerValidationExceptionSummary, \
-    ContainerValidationException
-from prestans3.types import Container, _Property, _PrestansTypeMeta
-from prestans3.utils import is_str, inject_class, with_metaclass, ImmutableMergingDictionary, LazyOneWayGraph, \
+from prestans3.errors import ValidationException, AccessError, ContainerValidationException
+from prestans3.types import Container, _Property
+from prestans3.utils import inject_class, with_metaclass, ImmutableMergingDictionary, LazyOneWayGraph, \
     _PrestansTypeMeta
