+++ prestans3/types/p_time.py
@@ -8,7 +8,7 @@
     :copyright: (c) 2016 Anomaly Software
     :license: Apache 2.0, see LICENSE for more details.
 """
-from datetime import time, tzinfo
+from datetime import time
 
 from prestans3.types.temporal import Temporal
