+++ prestans3/utils.py
@@ -8,12 +8,8 @@
     :copyright: (c) 2016 Anomaly Software
     :license: Apache 2.0, see LICENSE for more details.
 """
-import functools
-import numbers
-import sys
 from copy import copy
 
-import prestans3.types
 from prestans3.errors import AccessError
 
 try:
@@ -282,6 +278,7 @@ class LazyOneWayGraph(dict):
         self[of_type] = MergingProxyDictionary({}, *mro_)
         return self[of_type]
 
+
 class _PropertyRulesProperty(object):
     # noinspection PyUnusedLocal
     def __get__(self, cls, _mcs):
@@ -307,233 +304,3 @@ class _PrestansTypeMeta(type):
     _prepare_functions_graph = None
     _config_check_graph = None
     _property_rule_graph = None
-
-
-########################################################################################################################
-# The following code was obtained from the `future library`_ and is used in compliance with the `MIT licence`_         #
-#                                                                                                                      #
-# .. _future library: https://pypi.python.org/pypi/future                                                              #
-# .. _MIT Licence: https://opensource.org/licenses/MIT                                                                 #
-########################################################################################################################
-
-
-
-PY3 = sys.version_info[0] == 3
-PY2 = sys.version_info[0] == 2
-PY26 = sys.version_info[0:2] == (2, 6)
-PY27 = sys.version_info[0:2] == (2, 7)
-PYPY = hasattr(sys, 'pypy_translation_info')
-
-
-def istext(obj):
-    """
-    Deprecated. Use::
-        >>> isinstance(obj, str)
-    after this import:
-        >>> from future.builtins import str
-    """
-    return isinstance(obj, type(u''))
-
-
-def isbytes(obj):
-    """
-    Deprecated. Use::
-        >>> isinstance(obj, bytes)
-    after this import:
-        >>> from future.builtins import bytes
-    """
-    return isinstance(obj, type(b''))
-
-
-def isnewbytes(obj):
-    """
-    Equivalent to the result of ``isinstance(obj, newbytes)`` were
-    ``__instancecheck__`` not overridden on the newbytes subclass. In
-    other words, it is REALLY a newbytes instance, not a Py2 native str
-    object?
-    """
-    # TODO: generalize this so that it works with subclasses of newbytes
-    # Import is here to avoid circular imports:
-    from prestans3.future.newbytes import newbytes
-    return type(obj) == newbytes
-
-
-def isint(obj):
-    """
-    Deprecated. Tests whether an object is a Py3 ``int`` or either a Py2 ``int`` or
-    ``long``.
-
-    Instead of using this function, you can use:
-
-        >>> from future.builtins import int
-        >>> isinstance(obj, int)
-
-    The following idiom is equivalent:
-
-        >>> from numbers import Integral
-        >>> isinstance(obj, Integral)
-    """
-
-    return isinstance(obj, numbers.Integral)
-
-
-# Some utility functions to enforce strict type-separation of unicode str and
-# bytes:
-def disallow_types(argnums, disallowed_types):
-    """
-    A decorator that raises a TypeError if any of the given numbered
-    arguments is of the corresponding given type (e.g. bytes or unicode
-    string).
-
-    For example:
-
-        @disallow_types([0, 1], [unicode, bytes])
-        def f(a, b):
-            pass
-
-    raises a TypeError when f is called if a unicode object is passed as
-    `a` or a bytes object is passed as `b`.
-
-    This also skips over keyword arguments, so
-
-        @disallow_types([0, 1], [unicode, bytes])
-        def g(a, b=None):
-            pass
-
-    doesn't raise an exception if g is called with only one argument a,
-    e.g.:
-
-        g(b'Byte string')
-
-    Example use:
-
-    >>> class newbytes(object):
-    ...     @disallow_types([1], [unicode])
-    ...     def __add__(self, other):
-    ...          pass
-
-    >>> newbytes('1234') + u'1234'      #doctest: +IGNORE_EXCEPTION_DETAIL
-    Traceback (most recent call last):
-      ...
-    TypeError: can't concat 'bytes' to (unicode) str
-    """
-
-    def decorator(function):
-
-        @functools.wraps(function)
-        def wrapper(*args, **kwargs):
-            # These imports are just for this decorator, and are defined here
-            # to prevent circular imports:
-            from prestans3.future.newbytes import newbytes
-            from prestans3.future.newstr import newstr
-            from prestans3.future.newint import newint
-
-            errmsg = "argument can't be {0}"
-            for (argnum, mytype) in zip(argnums, disallowed_types):
-                # Handle the case where the type is passed as a string like 'newbytes'.
-                if isinstance(mytype, str) or isinstance(mytype, bytes):
-                    mytype = locals()[mytype]
-
-                # Only restrict kw args only if they are passed:
-                if len(args) <= argnum:
-                    break
-
-                # Here we use type() rather than isinstance() because
-                # __instancecheck__ is being overridden. E.g.
-                # isinstance(b'abc', newbytes) is True on Py2.
-                if type(args[argnum]) == mytype:
-                    raise TypeError(errmsg.format(mytype))
-
-            return function(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
-
-
-def no(mytype, argnums=(1,)):
-    """
-    A shortcut for the disallow_types decorator that disallows only one type
-    (in any position in argnums).
-
-    Example use:
-
-    >>> class newstr(object):
-    ...     @no('bytes')
-    ...     def __add__(self, other):
-    ...          pass
-
-    >>> newstr(u'1234') + b'1234'     #doctest: +IGNORE_EXCEPTION_DETAIL
-    Traceback (most recent call last):
-      ...
-    TypeError: argument can't be bytes
-
-    The object can also be passed directly, but passing the string helps
-    to prevent circular import problems.
-    """
-    if isinstance(argnums, numbers.Integral):
-        argnums = (argnums,)
-    disallowed_types = [mytype] * len(argnums)
-    return disallow_types(argnums, disallowed_types)
-
-
-def issubset(list1, list2):
-    """
-    Examples:
-
-    >>> issubset([], [65, 66, 67])
-    True
-    >>> issubset([65], [65, 66, 67])
-    True
-    >>> issubset([65, 66], [65, 66, 67])
-    True
-    >>> issubset([65, 67], [65, 66, 67])
-    False
-    """
-    n = len(list1)
-    for startpos in range(len(list2) - n + 1):
-        if list2[startpos:startpos + n] == list1:
-            return True
-    return False
-
-
-def native(obj):
-    """
-    On Py3, this is a no-op: native(obj) -> obj
-
-    On Py2, returns the corresponding native Py2 types that are
-    superclasses for backported objects from Py3:
-
-    >>> from builtins import str, bytes, int
-
-    >>> native(str(u'ABC'))
-    u'ABC'
-    >>> type(native(str(u'ABC')))
-    unicode
-
-    >>> native(bytes(b'ABC'))
-    b'ABC'
-    >>> type(native(bytes(b'ABC')))
-    bytes
-
-    >>> native(int(10**20))
-    100000000000000000000L
-    >>> type(native(int(10**20)))
-    long
-
-    Existing native types on Py2 will be returned unchanged:
-
-    >>> type(native(u'ABC'))
-    unicode
-    """
-    if hasattr(obj, '__native__'):
-        return obj.__native__()
-    else:
-        return obj
-
-########################################################################################################################
-# The above code was obtained from the `future library`_ and is used in compliance with the `MIT licence`_         #
-#                                                                                                                      #
-# .. _future library: https://pypi.python.org/pypi/future                                                              #
-# .. _MIT Licence: https://opensource.org/licenses/MIT                                                                 #
-########################################################################################################################
